create shop form validation happens on mint button
